src/file_io.c : Use intptr_t instead of long for return value of _get_osfhandle.

+++ src/file_io.c
@@ -862,7 +862,7 @@ psf_set_stdio (SF_PRIVATE *psf, int mode)
 /* USE_WINDOWS_API */ void
 psf_set_file (SF_PRIVATE *psf, int fd)
 {	HANDLE handle ;
-	long osfhandle ;
+	intptr_t osfhandle ;
 
 	osfhandle = _get_osfhandle (fd) ;
 	handle = (HANDLE) osfhandle ;
